src/requirements.txt
cryptography
jwt-auth
requests
